refactor(details): move path
