+++ packages/web/components/Technology/Details/Tables/Costs.js
