+++ packages/web/components/Technology/Details/Tables/Responsibles.js
