+++ packages/web/components/Technology/Details/Tables/index.js
